langgraph_backend/pyproject.toml
[project]
name = "langgraph-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "bs4>=0.0.2",
    "crawl4ai>=0.6.3",
    "fastapi>=0.115.12",
    "langchain[google-genai]>=0.3.25",
    "langgraph>=0.4.3",
    "python-dotenv>=1.1.0",
    "sse-starlette>=2.3.5",
    "uvicorn>=0.34.2",
]
